Recreate project with latest Flutter
