+++ .packages
@@ -1,0 +1,22 @@
+# Generated by pub on 2019-05-17 10:13:53.671827.
+async:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/async-2.1.0/lib/
+boolean_selector:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/boolean_selector-1.0.4/lib/
+charcode:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/charcode-1.1.2/lib/
+collection:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/collection-1.14.11/lib/
+flutter:file:///usr/local/flutter/packages/flutter/lib/
+flutter_test:file:///usr/local/flutter/packages/flutter_test/lib/
+matcher:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/matcher-0.12.5/lib/
+meta:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/meta-1.1.6/lib/
+path:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/path-1.6.2/lib/
+pedantic:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/pedantic-1.5.0/lib/
+quiver:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/quiver-2.0.2/lib/
+sky_engine:file:///usr/local/flutter/bin/cache/pkg/sky_engine/lib/
+source_span:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/source_span-1.5.5/lib/
+stack_trace:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/stack_trace-1.9.3/lib/
+stream_channel:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/stream_channel-2.0.0/lib/
+string_scanner:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/string_scanner-1.0.4/lib/
+term_glyph:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/term_glyph-1.1.0/lib/
+test_api:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/test_api-0.2.4/lib/
+typed_data:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/typed_data-1.1.6/lib/
+vector_math:file:///usr/local/flutter/.pub-cache/hosted/pub.dartlang.org/vector_math-2.0.8/lib/
+flutter_gallery_plugin:lib/
