+++ CHANGELOG.md
@@ -1,7 +1,3 @@
-## 0.1.0
-
-* Version Modified
-
 ## 0.0.1
 
-* Initial Release
+* TODO: Describe initial release.
