+++ android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'flutter_gallary_plugin'
+rootProject.name = 'flutter_gallery_plugin'
