+++ example/.metadata
@@ -4,5 +4,7 @@
 # This file should be version controlled and should not be manually edited.
 
 version:
-  revision: f37c235c32fc15babe6dc7b7bc2ee4387e5ecf92
-  channel: beta
+  revision: 7a4c33425ddd78c54aba07d86f3f9a4a0051769b
+  channel: stable
+
+project_type: app
