+++ example/README.md
@@ -1,70 +1,16 @@
-```dart
-import 'package:flutter/material.dart';
-import 'dart:io';
-import 'dart:async';
+# flutter_gallery_plugin_example
 
-import 'package:flutter/services.dart';
-import 'package:image_gallery/flutter_gallary_plugin.dart';
+Demonstrates how to use the flutter_gallery_plugin plugin.
 
-void main() => runApp(
-    new MyApp());
+## Getting Started
 
-class MyApp extends StatefulWidget {
-  @override
-  _MyAppState createState() => new _MyAppState();
-}
+This project is a starting point for a Flutter application.
 
-class _MyAppState extends State<MyApp> {
-  List<Object> allImage = new List();
+A few resources to get you started if this is your first Flutter project:
 
-  @override
-  void initState() {
-    super.initState();
-    loadImageList();
-  }
+- [Lab: Write your first Flutter app](https://flutter.dev/docs/get-started/codelab)
+- [Cookbook: Useful Flutter samples](https://flutter.dev/docs/cookbook)
 
-  Future<void> loadImageList() async {
-    List allImageTemp;
-      allImageTemp = await FlutterGallaryPlugin.getAllImages;
-
-
-    setState(() {
-      this.allImage = allImageTemp;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return new MaterialApp(
-      debugShowCheckedModeBanner: false,
-      home: new Scaffold(
-        appBar: new AppBar(
-          title: const Text('Image Gallery'),
-        ),
-        body: _buildGrid()
-        ,
-      ),
-    );
-  }
-
-  Widget _buildGrid() {
-    return GridView.extent(
-        maxCrossAxisExtent: 150.0,
-        // padding: const EdgeInsets.all(4.0),
-        mainAxisSpacing: 4.0,
-        crossAxisSpacing: 4.0,
-        children: _buildGridTileList(allImage.length));
-  }
-
-  List<Container> _buildGridTileList(int count) {
-
-    return List<Container>.generate(
-    count,
-    (int index) =>
-    Container(child: Image.file(File(allImage[index].toString()),
-    width: 96.0,
-    height: 96.0,
-    fit: BoxFit.contain,)));
-  }
-}
-```
+For help getting started with Flutter, view our 
+[online documentation](https://flutter.dev/docs), which offers tutorials, 
+samples, guidance on mobile development, and a full API reference.
