--- ios/Classes/FlutterGallaryPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface FlutterGallaryPlugin : NSObject<FlutterPlugin>
-@end
