+++ ios/Classes/FlutterGalleryPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface FlutterGalleryPlugin : NSObject<FlutterPlugin>
+@end
