+++ ios/Classes/SwiftFlutterGalleryPlugin.swift
@@ -2,43 +2,43 @@ import Flutter
 import UIKit
 import Photos
 
-public class SwiftFlutterGallaryPlugin: NSObject, FlutterPlugin {
-    public static func register(with registrar: FlutterPluginRegistrar) {
-        let channel = FlutterMethodChannel(name: "image_gallery", binaryMessenger: registrar.messenger())
-        let instance = SwiftFlutterGallaryPlugin()
-        registrar.addMethodCallDelegate(instance, channel: channel)
-    }
+public class SwiftFlutterGalleryPlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {
+    let channel = FlutterMethodChannel(name: "flutter_gallery_plugin", binaryMessenger: registrar.messenger())
+    let instance = SwiftFlutterGalleryPlugin()
+    registrar.addMethodCallDelegate(instance, channel: channel)
+  }
 
     public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
         if (call.method == "getPlatformVersion") {
             result("iOS " + UIDevice.current.systemVersion)
         }
         else if (call.method == "getAllImages") {
-            
+
             DispatchQueue.main.async {
-                
+
                 let imgManager = PHImageManager.default()
                 let requestOptions = PHImageRequestOptions()
                 requestOptions.isSynchronous = true
                 let fetchOptions = PHFetchOptions()
                 fetchOptions.sortDescriptors = [NSSortDescriptor(key:"creationDate", ascending: true)]
-                
+
                 let fetchResult = PHAsset.fetchAssets(with: PHAssetMediaType.image, options: fetchOptions)
                 var allImages = [String]()
-                
+
                 var totalItration = 0
                 print("fetchResult.count : \(fetchResult.count)")
-                
+
                 var savedLocalIdentifiers = [String]()
-                
+
                 for index in 0..<fetchResult.count
                 {
                     let asset = fetchResult.object(at: index) as PHAsset
                     let localIdentifier = asset.localIdentifier
                     savedLocalIdentifiers.append(localIdentifier)
-                    
+
                     imgManager.requestImage(for: asset, targetSize: CGSize(width: 512.0, height: 512.0), contentMode: PHImageContentMode.aspectFit, options: PHImageRequestOptions(), resultHandler:{(image, info) in
-                        
+
                         if image != nil {
                             var imageData: Data?
                             if let cgImage = image!.cgImage, cgImage.renderingIntent == .defaultIntent {
