+++ lib/flutter_gallery_plugin.dart
@@ -2,13 +2,13 @@ import 'dart:async';
 
 import 'package:flutter/services.dart';
 
-class FlutterGallaryPlugin {
-  static const MethodChannel _channel = const MethodChannel('image_gallery');
+class FlutterGalleryPlugin {
+  static const MethodChannel _channel =
+      const MethodChannel('flutter_gallery_plugin');
   static const String METHOD_GET_ALL = "getAllImages";
   static const String METHOD_GET_FOR_PERIOD = "getImagesForPeriod";
   static const String ARGUMENT_PERIOD_START = "startPeriod";
   static const String ARGUMENT_PERIOD_END = "endPeriod";
-
 
   static Future<List<String>> getAllImages() async {
     List<dynamic> list = await _channel.invokeMethod(METHOD_GET_ALL);
