+++ pubspec.yaml
@@ -1,17 +1,18 @@
-name: image_gallery
-description: Flutter plugin to showing all the images from the storage in Android and iOS .
-version: 0.1.0
-authors:
-- Manoj Kumar <manoj@aeologic.com>
-- Virander <virander@aeologic.com>
-homepage: http://www.flutterdevs.com
-documentation:
+name: flutter_gallery_plugin
+description: A new flutter plugin project.
+version: 0.0.1
+author:
+homepage:
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.1.0 <3.0.0"
 
 dependencies:
   flutter:
+    sdk: flutter
+
+dev_dependencies:
+  flutter_test:
     sdk: flutter
 
 # For information on the generic Dart part of this file, see the
@@ -19,9 +20,13 @@ dependencies:
 
 # The following section is specific to Flutter.
 flutter:
+  # This section identifies this Flutter project as a plugin project.
+  # The androidPackage and pluginClass identifiers should not ordinarily
+  # be modified. They are used by the tooling to maintain consistency when
+  # adding or updating assets for this project.
   plugin:
-    androidPackage: adhoc.successive.com.fluttergallaryplugin
-    pluginClass: FlutterGallaryPlugin
+    androidPackage: com.wowanders.flutter_gallery_plugin
+    pluginClass: FlutterGalleryPlugin
 
   # To add assets to your plugin package, add an assets section, like this:
   # assets:
@@ -29,10 +34,10 @@ flutter:
   #  - images/a_dot_ham.jpeg
   #
   # For details regarding assets in packages, see
-  # https://flutter.io/assets-and-images/#from-packages
+  # https://flutter.dev/assets-and-images/#from-packages
   #
   # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.io/assets-and-images/#resolution-aware.
+  # https://flutter.dev/assets-and-images/#resolution-aware.
 
   # To add custom fonts to your plugin package, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
@@ -52,4 +57,4 @@ flutter:
   #         weight: 700
   #
   # For details regarding fonts in packages, see
-  # https://flutter.io/custom-fonts/#from-packages
+  # https://flutter.dev/custom-fonts/#from-packages
